requirements.txt
aioredis==1.2.0
amqp==2.3.2
asgi-redis==1.4.3
asgiref==2.3.2
astroid==1.6.5
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.1.18
channels==2.1.6
channels-redis==2.3.2
chardet==3.0.4
constantly==15.1.0
daphne==2.2.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
django-timezone-field==3.0
django-toolbelt==0.0.1
django-websocket-redis==0.5.2
djangorestframework==3.8.2
future==0.16.0
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.8.1
hiredis==0.3.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.3.4
kombu==4.2.2.post1
msgpack==0.5.6
msgpack-python==0.5.6
nltk==3.4
oauthlib==2.0.7
psycopg2==2.7.5
PyHamcrest==1.9.0
pylint==1.9.2
PySocks==1.6.8
python-crontab==2.3.5
python-dateutil==2.7.5
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
singledispatch==3.4.0.3
six==1.11.0
static3==0.7.0
textblob==0.15.2
tweepy==3.7.0
Twisted==18.9.0
TwitterSearch==1.0.2
txaio==18.8.1
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
wrapt==1.10.11
